Added the TeachingTimetables and Updates modules
The Updates module extracts a list of hash values. Every hash
value is a content version indicator. This allows the mobile app
to distinguish between the new and old content. The mobile app will
then request for the content only if the new content is available.

The TeachingTimetables module uses the database for the static
content also available on the faculty sites. Having the content
inside the db is more robust. No html parsing needed.